es setup and docker support

+++ src/snap-core/Snap.cs
@@ -6,6 +6,8 @@ using System.Linq;
 using System.Security.AccessControl;
 using System.Security.Principal;
 using System.Text;
+using Docker.DotNet;
+using Docker.DotNet.Models;
 using EnvironmentBuilder;
 using EnvironmentBuilder.Abstractions;
 using EnvironmentBuilder.Extensions;
@@ -14,6 +16,7 @@ using Microsoft.Data.SqlClient;
 using Microsoft.Extensions.Configuration;
 using Microsoft.SqlServer.Management.Common;
 using Microsoft.SqlServer.Management.Smo;
+using Nest;
 
 namespace Snap.Core
 {
@@ -140,10 +143,14 @@ namespace Snap.Core
         {
             foreach (var target in _configuration.Targets.Where(t => t.Pack != null && t.Pack.Enable))
             {
-                switch (target.Type)
+                switch (target.Type.ToLower())
                 {
                     case "mssql":
-                        RunMssqlPack(target); break;
+                        RunMssqlPack(target);
+                        break;
+                    case "elasticsearch":
+                        RunElasticSearchPack(target);
+                        break;
                     default: LogAndThrow($"Unknown target type {target}"); break;
                 }
             }
@@ -161,7 +168,11 @@ namespace Snap.Core
                 sb.Append($"_{repo.Head.FriendlyName}");
             }
 
-            return sb.ToString();
+            return sb
+                .ToString()
+                .Replace('/', '_')
+                .Replace('\\', '_')
+                .Replace('.', '_');
         }
         private void RunMssqlPack(SnapConfiguration.Target target)
         {
@@ -183,7 +194,7 @@ namespace Snap.Core
             backup.BackupSetName = $"{connectionStringBuilder.InitialCatalog} Backup";
             backup.Database = connectionStringBuilder.InitialCatalog;
 
-            var targetUniqueName = $"{GetMssqlTargetUniqueName(target)}";
+            var targetUniqueName = $"{GetMssqlTargetUniqueName(target)}.bkp";
 
             // Declare a BackupDeviceItem by supplying the backup device file name in the constructor, and the type of device is a file.   
             BackupDeviceItem backupDeviceItem = new BackupDeviceItem(targetUniqueName, DeviceType.File);
@@ -208,10 +219,15 @@ namespace Snap.Core
 
             // Set the database recovery mode back to its original value.  
             db.RecoveryModel = (RecoveryModel)recoveryModel;
-            
-            Utilities.Move(Path.Join(srv.BackupDirectory, targetUniqueName),
-                Utilities.GetPathForCurrentUser(targetUniqueName));
-            Utilities.Remove(Path.Join(srv.BackupDirectory, targetUniqueName));
+
+            FileSystemUtils.Move(Path.Join(srv.BackupDirectory, targetUniqueName),
+                FileSystemUtils.GeneratePathForCurrentUser(targetUniqueName));
+            FileSystemUtils.Remove(Path.Join(srv.BackupDirectory, targetUniqueName));
+        }
+
+        private void RunElasticSearchPack(SnapConfiguration.Target target)
+        {
+            var elastic = new ElasticClient(new Uri(target.Properties["Host"]));
         }
 
         private void MssqlIncrementalBackupAndRestore()
@@ -362,8 +378,13 @@ namespace Snap.Core
             // Store the current recovery model in a variable.   
             int recoveryModel = (int)database.DatabaseOptions.RecoveryModel;
 
+            var targetUniqueName = $"{GetMssqlTargetUniqueName(target)}.bkp";
+
+            // TODO make this more reliable and account for mssql being run in docker
+            FileSystemUtils.Move(FileSystemUtils.GeneratePathForCurrentUser(targetUniqueName), Path.Join(srv.BackupDirectory));
+
             // Declare a BackupDeviceItem by supplying the backup device file name in the constructor, and the type of device is a file.   
-            BackupDeviceItem backupDeviceItem = new BackupDeviceItem(target.Type + ".bkp", DeviceType.File);
+            BackupDeviceItem backupDeviceItem = new BackupDeviceItem(targetUniqueName, DeviceType.File);
 
             // Delete the AdventureWorks2012 database before restoring it  
             database.Drop();
@@ -397,6 +418,13 @@ namespace Snap.Core
 
             // Set the database recovery mode back to its original value.  
             database.RecoveryModel = (RecoveryModel)recoveryModel;
+
+            FileSystemUtils.Remove(FileSystemUtils.GeneratePathForCurrentUser(targetUniqueName));
+        }
+
+        private void RunElasticvSearchUnpack(SnapConfiguration.Target target)
+        {
+
         }
 
         private void RunClean()
@@ -417,29 +445,18 @@ namespace Snap.Core
         public string ConfigurationFile { get; set; }
         public List<Target> Targets { get; set; }
         public Dictionary<string, string> Properties { get; set; } = new Dictionary<string, string>();
+
         public class Target
         {
             public string Type { get; set; }
+            public bool IsRunningInDocker { get; set; }
             public Dictionary<string, string> Properties { get; set; } = new Dictionary<string, string>();
-            public PackTask Pack { get; set; }
-            public UnpackTask Unpack { get; set; }
-            public CleanTask Clean { get; set; }
-
+            public CommonTask Pack { get; set; }
+            public CommonTask Unpack { get; set; }
+            public CommonTask Clean { get; set; }
         }
 
-        public class PackTask
-        {
-            public bool Enable { get; set; }
-            public string TargetDir { get; set; }
-        }
-
-        public class UnpackTask
-        {
-            public bool Enable { get; set; }
-            public string TargetDir { get; set; }
-        }
-
-        public class CleanTask
+        public class CommonTask
         {
             public bool Enable { get; set; }
         }
@@ -462,22 +479,79 @@ namespace Snap.Core
         }
     }
 
-    public static class Utilities
+    public static class FileSystemUtils
     {
-        public static void Remove(string from)
+        public static bool Remove(string from)
         {
             if (File.Exists(from))
             {
                 File.Delete(from);
+                return true;
             }
+
+            return false;
         }
-        public static void Move(string from, string to)
+
+        public enum FileSystemType
         {
-            File.Move(from, to, true);
+            Local,
+            Docker
         }
-        public static string GetPathForCurrentUser(string fileName = null)
+
+        public static bool Move(string from, string to)
+        {
+            if (File.Exists(from))
+            {
+                File.Move(from, to, true);
+                return true;
+            }
+
+            return false;
+        }
+
+        public static bool MoveVirtual(string from, FileSystemType fromType, string to, FileSystemType toType,
+            string containerId)
+        {
+            if (fromType == FileSystemType.Local && toType == FileSystemType.Local)
+            {
+                return Move(from, to);
+            }
+            if (fromType == FileSystemType.Docker && toType == FileSystemType.Docker)
+            {
+                throw new NotImplementedException("Moving files between docker containers is not supported");
+            }
+
+            var client = new DockerClientConfiguration(new Uri("npipe://./pipe/docker_engine")).CreateClient();
+            var container = client.Containers.ListContainersAsync(
+                    new ContainersListParameters())
+                .GetAwaiter().GetResult()
+                .Single(x => x.ID.Contains(containerId) || x.Names.Any(n => n.ToLower() == containerId.ToLower()));
+
+            if (fromType == FileSystemType.Docker)
+            {
+                var response = client.Containers.GetArchiveFromContainerAsync(container.ID, new GetArchiveFromContainerParameters { Path = from }, false).GetAwaiter().GetResult();
+                if (File.Exists(to))
+                    File.Delete(to);
+
+                using (var fileStream = File.Create(to))
+                {
+                    response.Stream.Seek(0, SeekOrigin.Begin);
+                    response.Stream.CopyTo(fileStream);
+                }
+            }
+            else
+            {
+                client.Containers
+                    .ExtractArchiveToContainerAsync(container.ID, new ContainerPathStatParameters { AllowOverwriteDirWithFile = true, Path = to },
+                        File.OpenRead(from)).GetAwaiter().GetResult();
+            }
+
+            return true;
+        }
+        public static string GeneratePathForCurrentUser(string fileName = null)
         {
             var rootPath = Path.Join(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "snap");
+
             if (!Directory.Exists(rootPath))
             {
                 Directory.CreateDirectory(rootPath);
@@ -487,11 +561,13 @@ namespace Snap.Core
             di.Attributes = FileAttributes.Normal;
 
             var ds = new DirectorySecurity(rootPath, AccessControlSections.Access);
+
             // Using this instead of the "Everyone" string means we work on non-English systems.
             SecurityIdentifier everyone = new SecurityIdentifier(WellKnownSidType.WorldSid, null);
             ds.AddAccessRule(new FileSystemAccessRule(everyone, FileSystemRights.FullControl, InheritanceFlags.ContainerInherit | InheritanceFlags.ObjectInherit, PropagationFlags.None, AccessControlType.Allow));
 
-            return rootPath + (!string.IsNullOrWhiteSpace(fileName) ? fileName : string.Empty);
+            return rootPath + (!string.IsNullOrWhiteSpace(fileName) ? $"\\{fileName}" : string.Empty);
         }
+
     }
 }
